pyproject.toml
[tool.poetry]
name = "dollarbot"
version = "0.1.0"
description = ""
authors = ["thevahidal <thevahidal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = "^22.0"
python-decouple = "^3.8"
requests = "^2.32.3"
persiantools = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
